requirements.txt
jpype1>=1.4.0
pandas>=1.5.0
matplotlib>=3.5.0
openpyxl>=3.0.0
